Change the OpenAPI Spec to deduplicate the CloudEvent schema
The callbacks use the CloudEvent standard, wrapping the payload with a cloudEvent envelope. Rather than duplicating the definition of this envelope, the envelope is now defined once and the payload will be oneOf.

To enable this, the example for each callback has been defined in the callBack definition, rather than the CloudEvent schema definition.